Move RecipeParser class into a function module
